feat(gulf_client): add factory for creating A2A and Genkit connectors
This commit introduces a factory function, `createGulfConnector`, that allows for the easy creation of different `GulfConnector` instances based on a `ServerType` enum. This makes it simple for the application to switch between the A2A and Genkit backends.

The example application has been updated to use this new factory, and a UI toggle has been added to allow for easy switching between the two server types for testing and demonstration purposes.